separated spec files for columns
